fix sparse vectorset recommendation and use sparse vector to compute distance

+++ src/main/java/com/guokr/simbase/store/VectorSet.java
@@ -4,7 +4,7 @@ import com.guokr.simbase.events.VectorSetListener;
 
 public interface VectorSet {
 
-    public void rescore(int srcVecId, float length, float[] vector, Recommendation rec);
+    public void rescore(int srcVecId, int length, int[] vector, Recommendation rec);
 
     public void remove(int vecid);
